+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/editor/timings/ExcelTimingWriter.java
@@ -70,35 +70,20 @@ import org.preesm.ui.scenario.editor.SaveAsWizard;
  */
 public class ExcelTimingWriter extends ExcelWriter {
 
-  /** The scenario. */
   private final Scenario scenario;
 
   /**
-   * Instantiates a new excel timing writer.
-   *
-   * @param scenario
-   *          the scenario
    */
   public ExcelTimingWriter(final Scenario scenario) {
     super();
     this.scenario = scenario;
   }
 
-  /*
-   * (non-Javadoc)
-   *
-   * @see org.eclipse.swt.events.SelectionListener#widgetDefaultSelected(org.eclipse.swt.events.SelectionEvent)
-   */
   @Override
   public void widgetDefaultSelected(final SelectionEvent e) {
     // no behavior by default
   }
 
-  /*
-   * (non-Javadoc)
-   *
-   * @see org.eclipse.swt.events.SelectionListener#widgetSelected(org.eclipse.swt.events.SelectionEvent)
-   */
   @Override
   public void widgetSelected(final SelectionEvent e) {