+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/wizards/NewScenarioFileWizard.java
@@ -55,22 +55,12 @@ public class NewScenarioFileWizard extends BasicNewFileResourceWizard {
   public static final String SCENARIO_INITIAL_CONTENT = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + "<scenario>\n"
       + "</scenario>";
 
-  /*
-   * (non-Javadoc)
-   *
-   * @see org.eclipse.ui.wizards.newresource.BasicNewFileResourceWizard#addPages()
-   */
   @Override
   public void addPages() {
     super.addPages();
     super.setWindowTitle("New Scenario File");
   }
 
-  /*
-   * (non-Javadoc)
-   *
-   * @see org.eclipse.ui.wizards.newresource.BasicNewFileResourceWizard#performFinish()
-   */
   @Override
   public boolean performFinish() {
     final WizardNewFileCreationPage page = (WizardNewFileCreationPage) (getPage("newFilePage1"));